Fixed various setup problems so all tests pass, and updated the README.

+++ bootstrap.sh
@@ -3,7 +3,7 @@
 #
 # Python
 #
-rm -rf build/
+rm -rf build/ .eggs/ _python-env openfda.egg-info/
 
 export LANG=C
 
@@ -13,20 +13,12 @@ export LANG=C
 PYTHON_ENV=$(python -c 'import os,sys;print os.path.realpath(sys.argv[1])' ./_python-env)
 
 PYTHON=$(which python)
-if [[ -n MACHTYPE ]]; then
-  PIP="pip2 install"
-else
-  PIP="pip2 install --user"
-fi
-
-sudo -H $PIP virtualenv
-
 # Setup virtualenv if it doesn't exist.
 test -e $PYTHON_ENV || virtualenv -p $PYTHON $PYTHON_ENV
+test -e $PYTHON_ENV || ( echo 'You need virtualenv!' && exit 1)
 
 # Install project sources and dependencies into the environment
 $PYTHON_ENV/bin/pip install awscli
-$PYTHON_ENV/bin/pip uninstall -y openfda || true
 $PYTHON_ENV/bin/pip install -U -r  requirements.txt
 $PYTHON_ENV/bin/python setup.py develop
